corrections of mmove and split

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: elbarry <elbarry@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 15:47:21 by elbarry           #+#    #+#             */
-/*   Updated: 2025/11/21 19:06:51 by elbarry          ###   ########.fr       */
+/*   Updated: 2025/11/21 19:46:43 by elbarry          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*cpy_src;
 	unsigned char	*cpy_dest;
